Start reorganizing ui gql queries and mutations
