+++ ui/src/graphql/index.ts
@@ -1,0 +1,10 @@
+export { GET_COLLECTION_DETAILS } from './queries/collection_details.ts';
+export { GET_COLUMN_DETAILS } from './queries/column_details.ts';
+export { GET_HOME_INFO } from './queries/session_available_collections.ts';
+export { GET_USERS } from './queries/users.ts';
+export { VALIDATE_COLUMN_DEFS } from './queries/validate_column_defs.ts';
+export { VALIDATE_COLLECTION_DEF } from './queries/validate_collection_defs.ts';
+
+export { UPDATE_COLLECTION } from './mutations/update_collection.ts';
+export { UPDATE_COLUMN_DEFINITION } from './mutations/update_column_defs.ts';
+export { CREATE_COLLECTION_INIT } from './mutations/create_collection_init.ts';
