+++ ui/src/pages/ColumnManagement.tsx
@@ -12,7 +12,7 @@ import { useSession } from 'src/components/auth/session.tsx';
 import ColumnManagementForm from 'src/components/ColumnManagementForm.tsx';
 import { Link } from 'src/components/Link.tsx';
 import { LoadingBlock } from 'src/components/Loading.tsx';
-import { GET_COLUMN_DETAILS } from 'src/graphql/queries.ts';
+import { GET_COLUMN_DETAILS } from 'src/graphql/index.ts';
 import type { User } from 'src/graphql/schema.d.ts';
 
 const ErrorDisplay = function ({
